services/voice/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
websockets==12.0
httpx==0.27.0
spacy==3.7.4
pydantic==2.6.0
python-dotenv==1.0.1
redis==5.0.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.29
python-multipart==0.0.9
numpy==1.26.4
